Add new icons for experimental layout options in Gutenberg
These icons are experimental, so let's keep them in an unmerged pullrequest for now, so no reason to merge.

This PR adds three new icons:

- layout-two-fold
- align-pull-left
- align-pull-right

layout-two-fold uses CSS grid and flexbox to create a layout that is reminiscent of columns, but is _not_ columns. It is more like a table layout, but is also not a table layout, as it has responsive built in. We might want to rename and/or redesign this as we experiment in the branch, so just another reason to keep it in PR for now.

align-pull-left and right are, as they sound, alternatives to the traditional float left and float right. But they are not floats, and text doesn't wrap around them. These place themselves outside of the text, but still in context of it.